% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persimon_app_helpers.R
\name{runSimulation}
\alias{runSimulation}
\title{Run Persimon Simulation with User Inputs}
\usage{
runSimulation(input)
}
\arguments{
\item{input}{A list of input values from the 'shiny' app.}
}
\value{
A list containing two data frames:
\item{results}{Data frame with proportions of rejecting the null hypothesis}
\item{success}{Data frame with counts of successful model runs}
}
\description{
This internal function executes the persimon simulation using parameters
provided by the user through the persimon_app 'shiny' app interface.
}
\details{
This function serves as a wrapper around the \code{persimon} function, translating
user inputs from the persimon_app 'shiny' app into the format required by
\code{persimon}. It uses helper functions from the \code{mmints} package to process
certain inputs.
}
\seealso{
\code{\link{persimon}} for the underlying simulation function.
}
\keyword{internal}
